Modification du type fileSize

+++ src/main/java/ch/heigvd/res/io/util/BufferedIOBenchmarkData.java
@@ -17,7 +17,7 @@ public class BufferedIOBenchmarkData implements IData {
     
     private Map<String, Object> values;
 
-    public BufferedIOBenchmarkData(String operation, IOStrategy strategy, int blockSize, int fileSize, long duration) {
+    public BufferedIOBenchmarkData(String operation, IOStrategy strategy, int blockSize, long fileSize, long duration) {
         values.put("operation", operation);
         values.put("strategy", strategy);
         values.put("blockSizeInBytes", blockSize);
